+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.3
 
 - Improve the docs in `README.md` to include config and all of the core features
+- Fix help text in CLI output for `--sql-dialect` and `--output-dir` options
 
 ## 1.0.2